move versions and constant value validation to `sbe_schema_validator`
